refactor filter-builder
