+++ libs/shared/src/lib/components/dashboard-filter/filter-builder-modal/filter-builder-modal.const.ts
@@ -1,0 +1,167 @@
+/** Default filter structure */
+export const DEFAULT_STRUCTURE = {
+  showQuestionNumbers: 'off',
+};
+
+/**
+ * Array containing the different types of questions.
+ * Commented types are not yet implemented.
+ */
+export const QUESTION_TYPES = [
+  'text',
+  'checkbox',
+  'radiogroup',
+  'dropdown',
+  'tagbox',
+  'comment',
+  // 'rating',
+  // 'ranking',
+  // 'imagepicker',
+  'boolean',
+  // 'image',
+  'html',
+  // 'signaturepad',
+  'expression',
+  // 'matrix',
+  // 'matrixdropdown',
+  // 'matrixdynamic',
+  // 'multipletext',
+  'panel',
+  'paneldynamic',
+];
+
+/**
+ * Allowed properties for a core question in a child form.
+ */
+export const CORE_QUESTION_ALLOWED_PROPERTIES = [
+  'name',
+  'title',
+  'size',
+  'min',
+  'max',
+  'minValueExpression',
+  'maxValueExpression',
+  'minErrorText',
+  'maxErrorText',
+  'step',
+  'maxLength',
+  'placeholder',
+  'dateMin',
+  'dateMax',
+  'description',
+  'hideNumber',
+  'width',
+  'maxWidth',
+  'minWidth',
+  'startWithNewLine',
+  'indent',
+  'page',
+  'titleLocation',
+  'descriptionLocation',
+  'state',
+  'defaultValue',
+  'defaultValueExpression',
+  'relatedName',
+  'addRecord',
+  'updateRecord',
+  'addTemplate',
+  'Search resource table',
+  'visible',
+  'choicesFromQuestion',
+  'choices',
+  'choicesFromQuestionMode',
+  'choicesOrder',
+  'choicesByUrl',
+  'hideIfChoicesEmpty',
+  'choicesVisibleIf',
+  'choicesEnableIf',
+  'readOnly',
+  'isRequired',
+  'placeHolder',
+  'enableIf',
+  'visibleIf',
+  'tooltip',
+  'referenceData',
+  'referenceDataDisplayField',
+  'isPrimitiveValue',
+  'referenceDataFilterFilterFromQuestion',
+  'referenceDataFilterForeignField',
+  'referenceDataFilterFilterCondition',
+  'referenceDataFilterLocalField',
+  'showSelectAllItem',
+  'showNoneItem',
+  'showClearButton',
+  'bindings',
+  'choicesMin',
+  'choicesMax',
+  'allowClear',
+  'autoGrow',
+  'labelTrue',
+  'labelFalse',
+  'valueTrue',
+  'valueFalse',
+  'valueName',
+  'inputType',
+  'html',
+  'resource',
+  'displayField',
+  'test service',
+  'addRecordText',
+  'updateRecordText',
+  'canSearch',
+  'autoSelectFirstOption',
+  'searchButtonText',
+  'canOnlyCreateRecords',
+  'alwaysCreateRecord',
+  'prefillWithCurrentRecord',
+  'selectQuestion',
+  'staticValue',
+  'filterBy',
+  'filterCondition',
+  'selectResourceText',
+  'gridFieldsSettings',
+  'customFilterEl',
+  'customFilter',
+  'newCreatedRecords',
+  'afterRecordCreation',
+  'afterRecordSelection',
+  'afterRecordDeselection',
+  'displayAsGrid',
+  'valueExpression',
+  'canDelete',
+  'history',
+  'convert',
+  'update',
+  'inlineEdition',
+  'export',
+  'canDeselectRecords',
+  'autoSaveChanges',
+  'prefillWithValues',
+  'gqlUrl',
+  'gqlQuery',
+  'gqlPath',
+  'gqlValueName',
+  'gqlTitleName',
+  'gqlVariableMapping',
+];
+
+/**
+ * Navigation tab properties (will be disabled).
+ */
+export const NAVIGATION_PROPERTIES = [
+  'showPreviewBeforeComplete',
+  'pagePrevText',
+  'pageNextText',
+  'completeText',
+  'previewText',
+  'editText',
+  'startSurveyText',
+  'showNavigationButtons',
+  'showPrevButton',
+  'firstPageIsStarted',
+  'goNextPageAutomatic',
+  'showProgressBar',
+  'progressBarType',
+  'questionsOnPageMode',
+  'showTOC',
+];
